requirements.txt
PyOpenGL==3.1.10
glfw==2.10.0
numpy==2.3.4
pyrr==0.10.3
PILLOW==12.0.0
aiofiles==25.1.0